refactor: address 0.56.2 thermo review
