+++ apps/desktop-tauri/src-tauri/src/commands/bridge/pace.rs
@@ -1,0 +1,12 @@
+pub(crate) fn stage_str(stage: codexbar::core::PaceStage) -> &'static str {
+    use codexbar::core::PaceStage;
+    match stage {
+        PaceStage::OnTrack => "on_track",
+        PaceStage::SlightlyAhead => "slightly_ahead",
+        PaceStage::Ahead => "ahead",
+        PaceStage::FarAhead => "far_ahead",
+        PaceStage::SlightlyBehind => "slightly_behind",
+        PaceStage::Behind => "behind",
+        PaceStage::FarBehind => "far_behind",
+    }
+}
