+++ apps/desktop-tauri/src-tauri/src/commands/tests.rs
@@ -835,16 +835,6 @@ fn local_opencodego_estimates_keep_quota_windows_but_drop_derived_pace() {
 }
 
 #[test]
-fn authoritative_opencodego_sources_still_allow_derived_pace() {
-    assert!(super::provider_allows_pace(ProviderId::OpenCodeGo, "api"));
-    assert!(super::provider_allows_pace(ProviderId::OpenCodeGo, "web"));
-    assert!(!super::provider_allows_pace(
-        ProviderId::OpenCodeGo,
-        " LOCAL ESTIMATE "
-    ));
-}
-
-#[test]
 fn provider_cache_is_fresh_inside_stale_window() {
     assert!(super::is_provider_cache_fresh(
         Some(std::time::Instant::now()),
@@ -921,6 +911,7 @@ fn provider_cache_upsert_replaces_existing_provider() {
         cost: None,
         wayfinder_usage: None,
         source_label: "CLI".to_string(),
+        pace_authoritative: true,
     };
     let mut first =
         ProviderUsageSnapshot::from_fetch_result(ProviderId::Codex, &metadata, &result, None);
@@ -944,6 +935,7 @@ fn provider_cache_prunes_disabled_providers() {
         cost: None,
         wayfinder_usage: None,
         source_label: "CLI".to_string(),
+        pace_authoritative: true,
     };
     let codex =
         ProviderUsageSnapshot::from_fetch_result(ProviderId::Codex, &metadata, &result, None);
@@ -974,6 +966,7 @@ fn hiding_codex_spark_rows_preserves_other_extra_usage() {
         cost: None,
         wayfinder_usage: None,
         source_label: "CLI".to_string(),
+        pace_authoritative: true,
     };
     let mut snapshot =
         ProviderUsageSnapshot::from_fetch_result(ProviderId::Codex, &metadata, &result, None);
@@ -1004,6 +997,7 @@ fn claude_transient_auth_failure_preserves_first_last_good_snapshot() {
         cost: None,
         wayfinder_usage: None,
         source_label: "OAuth".to_string(),
+        pace_authoritative: true,
     };
     let good =
         ProviderUsageSnapshot::from_fetch_result(ProviderId::Claude, &metadata, &result, None);
@@ -1034,6 +1028,7 @@ fn claude_repeated_auth_failure_surfaces_error() {
         cost: None,
         wayfinder_usage: None,
         source_label: "OAuth".to_string(),
+        pace_authoritative: true,
     };
     let good =
         ProviderUsageSnapshot::from_fetch_result(ProviderId::Claude, &metadata, &result, None);
@@ -1069,6 +1064,7 @@ fn claude_cli_parse_failure_keeps_last_good_every_time() {
         cost: None,
         wayfinder_usage: None,
         source_label: "CLI".to_string(),
+        pace_authoritative: true,
     };
     let good =
         ProviderUsageSnapshot::from_fetch_result(ProviderId::Claude, &metadata, &result, None);
@@ -1104,6 +1100,7 @@ fn claude_hard_credentials_missing_does_not_preserve_stale() {
         cost: None,
         wayfinder_usage: None,
         source_label: "OAuth".to_string(),
+        pace_authoritative: true,
     };
     let good =
         ProviderUsageSnapshot::from_fetch_result(ProviderId::Claude, &metadata, &result, None);
@@ -1254,6 +1251,7 @@ fn japanese_provider_snapshot_localizes_weekly_label() {
         cost: None,
         wayfinder_usage: None,
         source_label: "OAuth".to_string(),
+        pace_authoritative: true,
     };
 
     let snapshot =
@@ -1283,6 +1281,7 @@ fn japanese_provider_snapshot_localizes_pace_reserve_description() {
         cost: None,
         wayfinder_usage: None,
         source_label: "OAuth".to_string(),
+        pace_authoritative: true,
     };
 
     let snapshot =
