+++ rust/src/core/usage_snapshot.rs
@@ -408,6 +408,14 @@ pub struct ProviderFetchResult {
 
     /// Label describing the data source (e.g., "oauth", "web", "cli")
     pub source_label: String,
+
+    /// Whether quota data is authoritative enough for pace/run-out advice.
+    #[serde(default = "default_pace_authoritative")]
+    pub pace_authoritative: bool,
+}
+
+fn default_pace_authoritative() -> bool {
+    true
 }
 
 impl ProviderFetchResult {
@@ -418,7 +426,14 @@ impl ProviderFetchResult {
             cost: None,
             wayfinder_usage: None,
             source_label: source_label.into(),
+            pace_authoritative: true,
         }
+    }
+
+    /// Mark this result as unsuitable for derived pace/run-out advice.
+    pub fn with_non_authoritative_pace(mut self) -> Self {
+        self.pace_authoritative = false;
+        self
     }
 
     /// Builder pattern: set cost
@@ -437,6 +452,17 @@ impl ProviderFetchResult {
 #[cfg(test)]
 mod tests {
     use super::*;
+
+    #[test]
+    fn fetch_result_pace_authority_defaults_true_and_can_be_disabled() {
+        let usage = UsageSnapshot::new(RateWindow::new(25.0));
+        assert!(ProviderFetchResult::new(usage.clone(), "api").pace_authoritative);
+        assert!(
+            !ProviderFetchResult::new(usage, "local estimate")
+                .with_non_authoritative_pace()
+                .pace_authoritative
+        );
+    }
 
     #[test]
     fn cost_snapshot_ignores_non_finite_values() {
