+++ rust/src/providers/codex/weekly_reset.rs
@@ -93,68 +93,8 @@ pub(super) enum DelayedDecision {
     Discard,
 }
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum ResetDiagnosticReason {
-    CandidateCreated,
-    SourceNotExactOAuth,
-    MissingPreviousSnapshot,
-    MissingWeeklyWindow,
-    ResetThresholdMismatch,
-    InvalidResetBoundary,
-    InconsistentResetBoundary,
-    UnsupportedResetBoundary,
-    PlanMismatch,
-    MissingCreditInventory,
-    ChangedCreditInventory,
-    EvidenceVersionMismatch,
-    FutureCandidate,
-    ExpiredCandidate,
-    StaleObservation,
-    MinimumDelay,
-    ConfirmedObservation,
-    StoreUnavailable,
-    StoreRequested,
-}
-
-impl ResetDiagnosticReason {
-    const fn code(self) -> &'static str {
-        match self {
-            Self::CandidateCreated => "candidateCreated",
-            Self::SourceNotExactOAuth => "sourceNotExactOAuth",
-            Self::MissingPreviousSnapshot => "missingPreviousSnapshot",
-            Self::MissingWeeklyWindow => "missingWeeklyWindow",
-            Self::ResetThresholdMismatch => "resetThresholdMismatch",
-            Self::InvalidResetBoundary => "invalidResetBoundary",
-            Self::InconsistentResetBoundary => "inconsistentResetBoundary",
-            Self::UnsupportedResetBoundary => "unsupportedResetBoundary",
-            Self::PlanMismatch => "planMismatch",
-            Self::MissingCreditInventory => "missingCreditInventory",
-            Self::ChangedCreditInventory => "changedCreditInventory",
-            Self::EvidenceVersionMismatch => "evidenceVersionMismatch",
-            Self::FutureCandidate => "futureCandidate",
-            Self::ExpiredCandidate => "expiredCandidate",
-            Self::StaleObservation => "staleObservation",
-            Self::MinimumDelay => "minimumDelay",
-            Self::ConfirmedObservation => "confirmedObservation",
-            Self::StoreUnavailable => "storeUnavailable",
-            Self::StoreRequested => "storeRequested",
-        }
-    }
-}
-
-fn log_reset_diagnostic(
-    stage: &'static str,
-    decision: &'static str,
-    reason: ResetDiagnosticReason,
-) {
-    tracing::debug!(
-        target: "codex_weekly_reset",
-        stage,
-        decision,
-        reason = reason.code(),
-        "Codex weekly-reset decision"
-    );
-}
+mod diagnostics;
+use diagnostics::{ResetDiagnosticReason, log_reset_diagnostic};
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 enum ResetCreditEvidence {
