+++ rust/src/providers/opencodego/local.rs
@@ -115,6 +115,7 @@ impl LocalUsageSnapshot {
         // but it is not server-confirmed authority. Keep that distinction in
         // the data contract so CLI/React can present it without guessing.
         ProviderFetchResult::new(snap, super::LOCAL_ESTIMATE_SOURCE_LABEL)
+            .with_non_authoritative_pace()
     }
 }
